android: rename simple example code
