+++ examples/clients/mobile/android/PawsKey/.gitignore
