+++ examples/clients/mobile/android/PawsKey/app/.gitignore
