+++ examples/clients/mobile/android/PawsKey/app/proguard-rules.pro
