+++ examples/clients/mobile/android/PawsKey/app/src/main/res/drawable/ic_launcher_background.xml
