+++ examples/clients/mobile/android/PawsKey/app/src/main/res/mipmap-xhdpi/ic_launcher.webp
