+++ examples/clients/mobile/android/PawsKey/app/src/main/res/mipmap-xxxhdpi/ic_launcher.webp
