+++ examples/clients/mobile/android/PawsKey/app/src/main/res/values/colors.xml
