+++ examples/clients/mobile/android/PawsKey/app/src/main/res/xml/backup_rules.xml
