+++ examples/clients/mobile/android/PawsKey/app/src/main/res/xml/data_extraction_rules.xml
