+++ examples/clients/mobile/android/PawsKey/gradle.properties
