+++ examples/clients/mobile/android/PawsKey/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Tue Aug 27 14:25:41 CEST 2024
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-8.7-bin.zip
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
