+++ examples/clients/mobile/android/PawsKey/gradlew.bat
