--- examples/clients/mobile/android/keypaws/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Fri Aug 23 18:07:36 CEST 2024
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-8.7-bin.zip
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
